Merge WIP monomorphisation code
The goal is to monomorphise polymorphic types and function based on
sizes of type arguments and boxed-ness and avoid boxing integers,
booleans, and maybe in the future records and other types as well.

The base mono types are:

- I8
- I32
- I64
- Ptr

All polymorphic types are monomorphised based on applications of these.

The code is a mess and it does not work yet. Proper implementation
requires a lot of refactoring, which I will be working on next.