Update impl syntax and simplify interpreter
When implementing a trait, add `for` between the trait name and the
implementing type.

This simplifies a bunch of code as we no longer have to look up the type
constructor in `impl T[X]` to see whether it's a trait or a type.

Most importantly this means we no longer need to pass the types to the
interpreter.

With the limitations we have in the current version of (1) not allowing
a type constructor to implement a trait with different arguments (e.g.
`ToStr for Vec[I32]` and `ToStr for Vec[I64]` are not allowed together)
(2) not allowing a type to have the same method defined multiple times
(even if it's for different traits), this means interpreter can just
ignore traits.