Update variant pat and expr syntax:
Use tilde instead of backtick: backtick makes it difficult to paste code
into Markdown documents, and also in string interpolations where we also
use backticks.

String interpolations are less likely to be pasted as single-line
comments in Markdown, so I think it's OK to keep using backticks there.

We should probably use underscores in the future when we have a module
system and can hide names, without having to add underscore to indicate
that the symbol is private.

Issue: #44