Implement exceptions (#51)
Every function gets an exception type. When not specified, the default exception
type is `[..r]` (where `r` is fresh). There's no exception type inference yet.

Calling a function with exception type `exn` unifies `exn` with the current
function's exception type.

Standard library functions:

- `throw` converts a value into an exception:
  ```
  prim fn throw[a, r](exn: [..r]): {..r} a
  ```

- `try` converts exceptions thrown by the given closure into `Result.Err`, and
  return value into `Result.Ok`:
  ```
  prim fn try[a, exn, r](cb: Fn(): {..exn} a): {..r} Result[[..exn], a]
  ```

- `untry` converts `Result.Err` values into exceptions:
  ```
  fn untry[a, errs](res: Result[[..errs], a]): {..errs} a
      match res:
          Result.Ok(a): a
          Result.Err(err): throw(err)
  ```

The exception signature syntax `{...}` is the same as variant syntax, just with
braces instead of brackets. This is to avoid ambiguity in `Fn(): []` where it's
unclear whether `[]` is the exception type or return type.

Closes #11.