Allow more types in exception type position in return type syntax (#93)
Currently the exception type of a function type needs to be variant type, with
the special syntax `{...}`.

This was done to avoid LR(1) conflicts in the grammar.

However it's too restrictive, and also inconsistent, as the variant syntax
normally uses brackets.

With this commit we add a hack to handle uppercase identifiers followed by left
bracket specially in the parser. This allows eliminating the ambiguity in the
lalrpop grammar when having two types next to each other:

    // Unclear whether to shift '[' and parse type arguments, or reduce 'A' as
    // a type.
    'A' '[' ...

With this change, when there's no space between an uppercase identifier and a
left bracket, we combine them as one token before parsing. So in the case above
the parser is forced to parse type arguments.

After this change there's one more ambiguity when parsing two types next to
each other:

    'Fn' ... ':' 'Fn' ... ':' 'T1' 'T2'

Here we have two potential parses:

    'Fn' ... ':' ('Fn' ... ':' 'T1' 'T2')
    'Fn' ... ':' ('Fn' ... ':' 'T1') 'T2'

To fix this we don't allow function types in exception type position, which
rules out the second alternative.

Fixes #53.