Unify ret ty of the fun before checking args when checking call exprs
Unifying return type of the function in a call expression before
checking the arguments make the expected type in call arguments more
precise.

This allows inferring type of integer literals in the new test
`ReturnTypeExpectedType.fir`:

    f(): Option[(a: U32, b: U32)]
        match Bool.False:
            Bool.True: Option.Some((a = 1, b = 2))
            Bool.False: Option.None

Before this commit, when checking `Option.Some(...)`, we instantiate the
scheme of `Option.Some`, then check the arguments, then apply the
instantiated type to the argument types.

When checking the arguments, we pass the instantiated type's argument
types as expected types.

Because the instantiation is fresh (no other unification is done at this
point), this means that the arguments are often not precise enough to
generate good error messages in the arguments and allow inferring
integer types.

With this change, we unify the return type of the instantiation with the
expected type, `Option[(a: U32, b: U32)]` in the example. This makes the
argument type `(a: U32, b: U32)`, which we use as the expected type
when checking the record expression, allowing inferring the integer
types.

This change reveals a bug in the unification where we unify rows with
non-rows sometimes. To fix this, we add kinds to `Ty` nodes, and check
the kinds when linking a variable to a type.