Compiler: implement scanner, update tokens with loc info
Token locations have redundant information: from source name and byte
index we can derive line and column numbers.

However to keep rest of the pipeline simple for now we add line and
column numbers to tokens.

Scanner is a port of the interpreter's scanner. We'll change it later to
address known issues such as #113, #92, #32.