Merge Constr and ConstrSelect expr types
In prep for variants with named types (instead of labels), merge
`Constr` (for products) and `ConstrSelect` (for sums).

`ConstrSelect.ty` is not optional. When not available, the expr is for a
product constructor.

With named types in variants, we can then add a field `variant: bool` to
one type instead of two, and handle it the same way regardless of
whether `ty` is available or not.