Implement explicit type arguments (#123)
Explicit type arguments allow calling methods that are normally ambiguous at
call sites, e.g.:

```
trait Iterator[iter, item, exn]:
    ...

    # `item` is ambiguous at the call sites.
    count(self: iter): exn U32
        let n = 0u32
        for _: item in self:
            n += 1
        n

    # `item` and `exn` are ambiguous at call sites.
    sizeHint(self: iter): U32
        0
```

In this example, `item` would ideally be an associated type and not ambiguous.
However I'm not sure if `exn` should also be an associated type.

Regardless, type applications have other uses too: for making types explicit
for documentation or debugging purposes.

Current implementation does not allow omitting some of the arguments. This is
left as a future work.

Order of type parameters are currently a bit confusing or counter intuitive.
For example, in the test we have this:

```
main
    print(Option.None[U32])
    print(foo[[], U32]())
    print(T(x = ()).f[U32, [], Str]())

foo(): Option[t]
    Option.None[t]

type T[a]:
    x: ()

T.f(self: T[a]): Option[t]
    Option.None[t]
```

In functions, when a type parameter appears in return type position, the type
parameter for the exception row comes before the type parameter. This is
consistent with the front-end syntax where exception type appears before the
return type, but I think we probably want the exception type parameters to
always be the last. We should update this with the new syntax proposed in
https://github.com/fir-lang/fir/issues/113#issuecomment-2897924906, where
exception type comes after the return type.

It shouldn't be difficult to fix this later. For now, we just want to be able
to call `Iterator.count` and `Iterator.sizeHint`.