New indentation token generator (#136)
This changes function signature syntax slightly and implement a new scanner, as
described in #113.

The new scanner eliminates all backslashes. Backslash is no longer used
anywhere in the syntax now (only in escapes in string and char literals), and
can even be removed from the lexer.

Summary of the new scanner, as described in #113:

- We get two scanner instead of one: one for indentation sensitive parts of the
  syntax, one for indentation insensitive parts.

- These two scanners call each other as they enter and exit indentation
  sensitive and insensitive parts of the syntax.

- Indentation sensitive parts are: code in `{...}`, and code at the top level.

- Indentation insensitive parts are: code in `(...)` and `[...]`.

- Each scanner function maintains its own indentation stack. This makes sure
  indentation in a delimited code cannot influence indentation of the wrapping
  code.

- In `{...}`, we no longer generate `NEWLINE; INDENT` and `DEDENT` at the end.
  Tokens start with the first token after `{`.

- In indentation sensitive scanner, we only generate `NEWLINE; INDENT` after a
  `:`.

  When not followed by a `:`, indented code becomes part of the previous line.

  Code at the same column as the previous generates a `NEWLINE`.

  Dedents generate `DEDENT` tokens as before.

- Indentation insensitive scanner is the same as before.

Because we only generate `NEWLINE; INDENT` after a `:`, function signatures are
now

    myFunction(arg1: Type1) exn ReturnType:
        ...

instead of

    myFunction(arg1: Type1): exn ReturnType
        ...

This makes the syntax more consistent as all indented code starts after a `:`
now.

Compiler's scanner is not updated yet.

Fixes #113.