Update function expression syntax
Use `||: <expr>` (and with indented form) instead of `{ <expr> }`, or
`fn() ReturnType / ExnType { <expr> }`.

The short form is still as short as the short form before, but the new
syntax is more consistent with the other syntax:

- Indented blocks now starts with `:` everywhere, no exceptions.
- Braces are no longer used. They're now removed from the parser
  entirely, parser no longer recognizes braces.

Removing braces from the grammar allows treating braces as `INDENT` and
`DEDENT` tokens in the scanner, potentially allowing indentation
insensitive parsing of all indentation sensitive syntax in the language,
without special handling in the parser. We'll explore this option later.

Fixes #148.