Implement syntax for iterators, call `fromIter` implicitly to use it for collection literals (#125)
This implements new syntax `[...]` for generating iterators.

The type checker implicitly converts `[...]` to `Type.fromIter([...])` when the expected type has a `fromIter` method.

When the expected type is not clear, `Type.[...]` syntax can be used to specify the type.

Not in this PR: the `Type` part cannot have type parameters, e.g. `Vec[U32].[...]` is not allowed yet. This will be implemented in an future PR.

This allows using this syntax for generating `Vec`s, `HashMap`s etc.

Example:

```
let x: HashMap[Str, U32] = [
    "a" = u32(1),
    "b" = u32(2),
]
print(x)
```

Fixes #118.