Add a mode to the interpreter to dump tokens, compare interpreter's tokens with compiler's
They currently generate different tokens. The test is not added to the
CI yet.

The differences seem to be because we haven't updated the compiler's
lexer after some of the recent changes (like removal of `u32` etc.
suffixes in integer literals), and because the interpreter's lexer
combines some of the tokens to help with LR(1) parsing.