Update array representation to allow slicing, implement `Array.slice`
`Array` is now a pointer + length. It's immutable, so when we have value
types we will make `Array` a value type.

This representation allows slicing by incrementing the pointer field and
decrementing the length field.

With this, `Str` becomes a "newtype" for `Array`. It will also be a
value type when we have value types.

Fixes #141.