Compiler: update scanner
Not tested on the CI yet: compiler's scanner generates NEWLINE tokens
with different locations. Not sure whether the bug is in the compiler's
scanner or the interpreter's scanner yet. Needs more debugging.

For now, the token kinds are right so we can proceed to parsing.

Scanner unit tests are removed as we have good test coverage now via
scan.sh. We should bring them back before deleting the interpreter.