New constructor declaration syntax (#183)
For consistency with the constructor call and record type syntax, and to allow
positional arguments in constructors, use parenthesized list of fields in
constructor declarations.

`type Foo` becomes a unit type (from void).

For now, `type Foo` and `type Foo()` are the same thing. Sum constructors can
also have an empty field list `()`.

In the future, we may want to allow only one of these.

Users can't define void types anymore. I'm not sure how useful those are, and
we already have a void type which is the empty variant `[]`.

Closes #66.
Closes #149.