Add tokens to AST nodes (#187)
Keep all tokens in a `Vec` and refer to them by index.

Token indices will be used to get AST locations (for errors) and to get
comments before/after a node when formatting.

In the compiled compiler token types will be value types, so the whole vector
will be two heap objects (vector + backing array) with no pointers.

See #135 for the discussion.