Add Clone trait, implement it for iterators
To avoid implementing our own iterator type (which would require a
shared runtime library for PEGs) in #188 (and because we want it
anyway), this adds a `Clone` trait to the standard library and
implements it for the iterator types, `Result`, and `Option`.

We don't implement it for iterator types that have a closure field. I'm
not sure how to implement traits for closures that capture values yet.

We also don't implement it for the containers yet as I'm also not sure
how to allocate the backing arrays when clonning. E.g. should we
compact, or should we allocate the same size?

For now we implement it for the types needed to unblock #188.