PEG: Use a trait to skip trivia in generated parsers (#191)
The new trait `TriviaTokens` is used to skip trivia in the scanner and
generated parsers:

    trait TriviaTokens[t]:
        isTrivia(self: t) Bool

Generated parser interface is also updated to to use a new state type:

    type ParserState[t](
        _tokens: Vec[t],
        _cursor: U32,
        _errorCursor: Option[U32],
    )


With the state type we no longer need to pass around cursor and error cursor
values, and we can also have a single `ParserState` method to skip trivia
tokens when checking the next token, which is done in `ParserState.peek`.

This design is not perfect: we don't have a way to *not* skip trivia tokens.

However it should be good enough to unblock us for now and make progress on the
formatter and compiler.

Closes #188.