Compiler: refactor scanner
Simplify the `scanIndented` interface by taking just the block column as
argument (instead of location of the first token).

Simplify `scanIndented` code by removing the `lastLoc` local and always
using the last token from `tokens` as the last generated token.

This will also help when generating trivia tokens, in code like:

    do:         # hi
        stmt1
        stmt2

Here we want to generate `NEWLINE` after `# hi`. In the future commits
we'll push and skip trivia tokens with `peek` and `next`. With the last
location always obtained from `newTokens` we won't need to update any
locals as we skip trivia tokens.