Compiler: re-associate binary expressions based on operator precedences
Parser generates a list of expressions with operators in between.

New function `associate` processes this list and creates `BinOp`
expressions based on operator precedences and associativities.

Operators that don't associate cause parser errors. Errors are currently
not helpful, we'll improve them later. Currently all Fir code is also
parsed by the Rust implementation, which gives somewhat helpful errors
(at least shows error locations), so error message quality in the
compiler is not that important yet.

(Compiler's parser is more strict as some operators don't commute in the
compiler, but it's already quite rare to combine e.g. `*` and `<<` in
the same expression without parentheses, so it should be OK for now.)

This is mainly tested in #196. Because we don't have a way to print ASTs
yet and constructing ASTs by hand is also difficult, this commit doesn't
come with tests. We'll address these issues later and add tests.

Operator precedences are copied from Carbon.

Fixes #119.