Formatter: pass tokens as array instead of vector
Using an array instead of vector eliminates a layer of indirection when
accessing the tokens.

Interpreter is slow so an extra layer of indirection does not matter
much currently (formatting the whole repo takes the same amount of
time), but it's still a good practice, and it'll make a difference when
we compile it.