Rename directories to make the project root a Fir library
In prep for the module system in #195, rename directories and update
`import`s so that the project root looks like a library.

With this we get rid of the import path flags and `FIR_ROOT` env
variable. `fir` (the interpreter) always assumes that the working
directory is the library root, and searches for files starting from the
working directory.

We don't have a concept of a package yet. Everything is one library,
and the library root is the working directory.

When running Fir files outside of the project root, we symlink the `Fir`
directory to the project's root, so that the implicit `import
Fir.Prelude` can be resolved.

Compiler unit tests start to fail with this commit as compiler doesn't
tokenize `Foo.Bar.Baz` as a single token, and `_combineTokens` (used to
compare generated tokens with the interpreter's) only combines two
components: `Foo.Bar`. We'll update `_combineTokens` later.