Update path syntax
Use slash instead of dot between path components, see [1] for details.

We don't support paths in expressions and types yet, only import
declarations are updated for now. Paths in expressions and types will be
implemented in the compiled version as it's too much work to update the
interpreter now, and the interpreter's parser doesn't like separated
things in expression context (LR(1) issues).

[1]: https://github.com/fir-lang/fir/issues/195#issuecomment-3285367461