Compiler: implement name resolving pass (#224)
This implements some of the plan in #195.

In the final compiler we may want to do name resolving during type checking,
but it's easier to debug a separate pass, so for now we do it separately.

High-level changes:

- A new type `Program` holds all of the modules.

- New identifier types introduced for identifiers that can refer to things in
  other modules.

- Identifiers start as "unresolved", name resolving pass updates them with the
  referred thing's (function, type, ...) module and the name in the module.

  (Reminder: we don't allow importing with renaming, but we will in the
  future.)

- A new `Visitor` trait added to by default visit every field of the visited
  AST type.

  For now it doesn't really visit all of the fields, we'll update as we need to
  visit more fields.

  Name resolver uses the visitor to traverse the AST.

- `ScopeMap` type copied from the interpreter to handle scoping when resolving
  term variables.

Name resolving works by first collecting the imports, and transitively loading
imported modules. It then creates strongly connected components (SCCs) of
module imports. Because an import currently imports everything in a module (and
without renaming), each SCC has access to the same definitions, so an
environment is created for each SCC, by first creating environments for
imported SCCs and then taking the union of all of those environments + the
current SCCs environment. Name conflicts are reported as errors.

After bootstrapping we'll allow importing specified identifiers, and with
different names. The code should be mostly ready to handle these features, for
example resolved ids can have different names locally and in the defining
module.