Refactor `HashMap.fromIter`, use `Array` as `HashMap` storage
Using `Iterator.size` is too fragile as combining two iterators with
known size can create an iterator with unknown size, and in general we
don't know in compile time whether an `Iterator` argument has known
size.

In the future we should have another trait, maybe `KnownSizeIterator`,
with a `size` method that returns the exact number.

For now, just take a size argument in `HashMap.fromIter` and panic if
the iterator doesn't yield that many elements.

With this we also start to use an `Array` (instead of a `Vec`) as
backing store of `HashMap`.