Don't print interpreter stack traces by default
When we added `--no-backtrace` the interpreter was buggy and being
developed, so a type error (or other crash) would often mean a bug in
the interpreter.

It's gotten quite stable since then, and most errors are actually
problems in Fir code. So update the default to not print the interpreter
backtrace. Turn the flag into `--backtrace` to opt-in.