Remove `{Result,Option}.{isSome,isNone}`
With `is` expressions they're no longer as useful, and they introduce an
extra way of doing the same thing without any obvious tradeoffs, making
it difficult to choose one and causing inconsistencies.

We also don't want to be adding `isFoo`, `isBar`, ... for every
constructor for consistency. Just use `is` always and remove these
functions.