Compiler: Start implementing the type checker (#228)
Compiler: start implementing the type checker

This commit contains two months of work implementing some parts of the type
checker:

- The data structures for type checking types, type constructors, the trait
  environment etc.

- Unification, conversions (AST types to type checking types), normalization,
  kind inference, and a few other basic functions.

We also fix/improve name resolving and visitors on the way.

The high-level idea is: we start loading files from the entry point and create
a dependency graph with the parsed import lists. (currently we can't parse just
the import lists so we parse whole files)

The dependency graph should then be processed in parallel and also
incrementally:

- `Scheme`s depend on `TyCon`s, `TyCon`s depend on other `TyCon`s.

- When initializing `TyCon`s, we process the sibling nodes in the dependency
  tree of SCCs in parallel.

- When a module's `TyCon`s are fully initialized, we can create its type
  schemes in parallel with everything else.

  Because every function has full type signature this can be done right after
  initializing the `TyCon`s. (no need to look at the code, no need to know
  imported schemes etc.)

Currently we don't have a way to do parallel computation in Fir, but the
current design should make it easy to process the modules in parallel as
described here.

This code can currently create per-module type checking environments for all of
the modules in the compiler + the standard libraries. (everything reachable
from `Compiler/Main.fir`)

Other than running to completion the code is not tested yet.

This code will probably need some refactoring but then we can continue with:

- Type checking function bodies.
- Resolving predicates.