Interpreter: collect generated predicates in a `Vec` instead of `Set`
See #229 for the discussion.

It's good to do this now as we're porting the type checker to Fir, and
in Fir we don't have an ordered set type yet. It's good to make the
refactoring in the Rust implementation first so that we won't have any
surprises in the Fir port (bugs, missing `Pred` impls etc.).

Closes #229.