Compiler: only generate TildeUpperId tokens when testing
We don't need combined tokens in the compiler, and this token is making
things more complicated in a few places in the compiler as the '~'
character is getting included in the token text, making it difficult to
get the type and constructor names in e.g. `~Bool.True`.

Remove `visitConstructor` in the name resolver: it was a hack to avoid
visiting `~Foo` as a type (rather than `Foo`), and it was buggy because
omitting variants meant `Foo` in `~Foo` wasn't resolved.