New closure syntax (#245)
Change closure argument list syntax from `|...|` to `\(...)`.

The previous syntax was confusing the scanner and it was only accidentally
working. The scanner expects commas to be in delimited blocks (parens, braces,
brackets), and pipes cannot be considered as delimited blocks because the
scanner doesn't have a way of distinguishing a binary operator pipe from a pipe
starting a delimited block.

The reason why closures still worked most of the time is because most of the
time they appeared in `map` and similar function arguments, so they were
already in between parens. But even then the token trees were not quite right,
e.g. in `map(|x, y|: ...)` the pipe and 'x' were considered one node of the
tree for the parenthesized part, and 'y' and the other pipe was another node.

With  the new syntax scanner just does the right thing and we can now have
closures with multiple arguments outside of any delimiters. The new test
`tests/ClosureMultipleArgsSyntax.fir` demonstrates the syntax that didn't parse
before.

Finally, the syntax with pipes always felt out of place in this language, and
we even had an issue about it: #161. The new syntax looks much better.

In the future we may want to omit the empty argument list: `\: ...` instead of
`\(): ...`.

Fixes #161.