Rewrite Format.sh, support formatting files outside of the repo
The previous version was somewhat over-engineered (created formatted
files in a temp directory, copied with `rsync`) and it didn't allow
formatting files outside of the Fir root.

Rewrite it to create formatted files next to the original files and then
`mv` it to the original locations, one file at a time.