Interpreter: fix/improve variants implementation (#251)
Generalize variant expression and pattern syntax from `~ <constr>` to
`~ <expr>` and `~ <pat>`.

This allows expressions and patterns like `~x` and `~"hi"` and simplifies
parsing.

We still require the type of the value being made a variant to be known. The
type constructor is still used as the label. If the type constructor is not
known we raise a type error.

Integer types are not allowed in variants yet as the interpreter cannot deal
with unboxed values in variants. This restriction will be lifted in the
compiler.