Compiler: pass module tokens to `{first,last}Token`
This allows searching in module tokens to find the first and last tokens
of AST nodes, without storing the tokens in the AST nodes.

We already find first/last tokens in the formatter without storing them
in the ASTs, but we need AST node locations in more places now (mainly
in the type checker), so it makes sense to do it in the AST API and
use it everywhere, instead of duplicating code.

This also refactors tokens type from `Vec[Token]` to `Array[Token]`.

This should allow removing more of the existing token fields, see #206.