Rename a few common names in the whole repo:
Both in the interpreter and compiler, use

- "Con" for constructors
- "Pat" for patterns
- "Sel" for selection (constructor, field, etc.)

We started with longer names but after two years of typing them a
million of times, I started to prefer shorter names. So over time they
became inconsistent with both longer (e.g. "constr") and shorter
versions in the same code. Rename the whole repo to consistently use the
shorter names.