Fix string interpolations (#255)
Current string interpolation handling in the interpreter is hacky and it only
handles on level of nesting.

In the compiler, implementing it in a hacky way is more difficult than doing it
properly. So we properly implement it in the interpreter as well as in the
compiler. The plan is mostly as described in #253: we generate `BEGIN_STR` and
`END_STR` tokens for strings, between these two tokens we generate
`BEGIN_INTERPOLATION` and `END_INTERPOLATION`. The lexer scans the
interpolation tokens in the initial state, so that allows nesting of strings
and comments in interpolations.

Scanner treats interpolation tokens as any other delimiter.
`START_INTERPOLATION` starts a non-indented block terminated by
`END_INTERPOLATION`.

Parser parses the interpolation tokens as an expression.

Formatting is currently not ideal: currently a string literal can only be split
into multiple lines in interpolations. That leads to formatting like: a string
literal going all the way to column 100 or more, and then the interpolated
expression is getting split at weird places because it has no more space left.

For now we accept this formatting: in the worst case the interpolated
expressions can be named in a let expression, making the literal both shorter
and better formatted.

In the future: we'll probably want a way to split a long string into multiple
strings. Then we can either manually split long strings, or maybe even find a
way for the formatter to do it. This is tracked in #256.

Fixes #87
Fixes #253