Interpreter: minor rewording and refactoring in monomorph
We no longer use the `r` type parameter for the exception type row,
instead we make functions parametric on the exception type (which may or
may not be a variant) and call it `exn`.

Also make it clear that `trait_ty_arg` is actually a type arg in an
`impl` block that we're matching against.