Lowering: return types of lowered expressions and statements (#257)
In prep for #244, refactor the lowering pass to return types of lowered
expressions and statements.

This will allow a more statically typed interpretation and compilation: instead
of checking fields of objects being indexed in runtime we can now generate
lowered indexing expressions with the offset of the fields being indexed.