Interpreter: rename `TyVar` as `UVar` for consistency with the compiler
In the interpreter we called unification variables `TyVar`s. It was
always confusing so in the compiler we called them `UVar`s, for
"unification variables". Rename `TyVar` as `UVar` in the interpreter.

There's still no constructor for "rigid type variables" in the
interpreter and those are still represented as `Ty::Con`s. We may
refactor this later if needed.