Fix/improve integer parsing, add min/max constants for integer types (#258)
Allow the whole U64 range in integer literals. Based on the inferred type of a
literal, check that the integer value is within the limits of the inferred
integer type.

New standard library functions `min` and `max` are added for all integer types.

`ToStr[I64]` implemented to be able to print the min/max values in the test.